+++ Phoenix/PHNotificationHelper.m
@@ -2,25 +2,35 @@
  * Phoenix is released under the MIT License. Refer to https://github.com/kasper/phoenix/blob/master/LICENSE.md
  */
 
+@import UserNotifications;
+
 #import "PHNotificationHelper.h"
 
 @implementation PHNotificationHelper
 
+static NSString *const PHNotificationHelperIdentifier = @"org.khirviko.Phoenix";
+
 #pragma mark - Delivering
 
-+ (void)deliver:(NSString *)message withDelegate:(id<NSUserNotificationCenterDelegate>)delegate {
-    NSUserNotificationCenter *center = [NSUserNotificationCenter defaultUserNotificationCenter];
-    center.delegate = delegate;
++ (void)deliver:(NSString *)message categoryIdentifier:(NSString *)identifier {
+    UNUserNotificationCenter *center = [UNUserNotificationCenter currentNotificationCenter];
 
-    NSUserNotification *notification = [[NSUserNotification alloc] init];
-    notification.informativeText = message;
-    notification.hasActionButton = NO;
+    UNMutableNotificationContent *content = [[UNMutableNotificationContent alloc] init];
+    content.body = message;
 
-    if (delegate) {
-        notification.hasActionButton = YES;
+    if (identifier) {
+        content.categoryIdentifier = identifier;
     }
 
-    [center deliverNotification:notification];
+    UNNotificationRequest *request = [UNNotificationRequest requestWithIdentifier:PHNotificationHelperIdentifier
+                                                                          content:content
+                                                                          trigger:nil];
+    [center addNotificationRequest:request
+             withCompletionHandler:^(NSError *error) {
+                 if (error) {
+                     NSLog(@"Error: Could not deliver notification. (%@)", error);
+                 }
+             }];
 }
 
 @end