+++ CHANGELOG.md
@@ -8,7 +8,14 @@ Release: dd.mm.yyyy
 
 ### Changes
 
+- **Breaking:** Function `App#launch(...)` is now asynchronous and the function does not return the app. Pass a callback function as the third argument to receive the app after launching.
 - Xcode 16 is now required for building.
+
+### API
+
+#### App
+
+- Change: Function `launch(...)` is now asynchronous and the function does not return the app. Pass a callback function as the third argument to receive the app after launching.
 
 4.0.1
 -----